Fix darwin support.
This was tricky; warnings about -pie not being used in compliation
caused Cairo's link test for pthreads to fail even though the link
worked.

I guess it found output on stderr and assumed it was  a big issue.

The fix, for posterity was:

export LDFLAGS=${LDFLAGS_CC}

We also --enable-threads since that catches the error at configure
time intead of getting more obscure compilation failures.